Switch to an ASGI compliant server for remote websockets (#881)
Closes #880, #595

This PR switches Sematic from gunicorn (a WSGI server) to uvicorn (an
ASGI server). This allows us to properly support websockets without
relying on monkey-patching from gevent/eventlet. As an added benefit, it
likely increases the load that the SocketIO server can handle, as async
handling is generally more performant than process/threading based
handling for websockets.

One of the main challenges in implementing this was that we need to
support two different configurations:

1. Sematic API (which uses Flask, a synchronous framework) and SocketIO
in one server, for local execution
2. Sematic API and SocketIO in different servers

If you use an ASGI server, in ASGI mode for (1), then you cannot use
Flask. However, if you use an ASGI server in WSGI mode for the SocketIO
server in (2), and try to run *any non-async framework alongside it in
the same server* you very quickly (i.e. even in minimal load testing)
run into a situation where the sync request handling (ex: for posts to
broadcast) stalls the event loop thread and throttles the websockets
noticeably. The solution was to:

1. For local mode, run the ASGI server in WSGI mode, and have SocketIO
use threading as its async mode. This seems to make SocketIO fall back
to polling transport, but that's not a big deal for local development.
2. Remotely, run the server in WSGI mode for the API, but ASGI mode for
SocketIO. In order to enable the SocketIO server to still respond to
HTTP posts about broadcasts it needs to make (and respond to liveness
probes), we needed to find a simple async framework. Starlette is a good
fit for this purpose. Note: though quart is closer to flask than
Starlette, it is less common and is currently breaking when bazel loads
its wheel. Therefore we opt for Starlette and NOT quart.

At some point in the future, we might want to switch to FastAPI or
Starlette entirely, which would simplify the above bifurcation. However,
that will require non-trivial work on the rest of the app.

Testing
-------
### Local

#### `bazel run //sematic/cli:main -- start`

- confirmed that it starts a server on the correct port
- ran examples against it
- confirmed that starting/stopping it repeatedly worked
- built and locally installed a fresh wheel, cleared my Sematic env, and
confirmed that it could run an example
- confirmed multiple starts in a row showed the "Sematic is already
running" message
- confirmed that closing the shell where the server was started from
left it running
- confirmed event updates are reflected in the UI
- confirmed the local log files are still being generated
(~/.sematic/*.log)
- confirmed that output from the server is hidden from stdout

#### `bazel run //sematic/api:server -- --debug`

- confirmed it starts server on the correct port
- ran examples against it
- confirmed event updates are reflected in the UI
- confirmed output from the server is shown in stdout

### Remote

- Tried posting to `/api/events/pipeline/update` on the deployed server,
with and without an API key (to confirm auth is working for Starlette)
- ran the testing pipeline against it
- confirmed that updates are reflected in the UI
- confirmed the browser can use the websocket transport
- ran the test pipeline with the images setting, to test CORs stuff

---------

Co-authored-by: Josh Bauer <josh@sematic.dev>