Upgrade pytorch to 1.13.1 (#894)
Closes #622

**Important**: This only impacts the version of pytorch that will be
used for developing Sematic itself. Users have and will continue to be
free to select pytorch versions higher or lower than this for their own
repos.

---------

When torch moved from 1.12.x to 1.13.0, it changed the way it depended
on some Cuda .so files. Specifically, instead of bundling these .so
files with the torch wheel, torch depends on some nvidia wheels that
contain the .so files. There are two ways that torch goes about trying
to find those .so files:

(1) Via the standard OS linking process. So when a torch .so is loaded,
the OS's linker is used to locate the .so files the torch one depends
on.
(2) If the first mechanism fails, it falls back to [a
hack](https://github.com/pytorch/pytorch/blob/1a6ab8a5dcce45789a6dd61261bc73428439ddb8/torch/__init__.py#L196
) to find & load the .so s from nvidia python libs

So if your base image/host env *already contain the required Cuda libs*
without installing any python dependencies and the linker path is
configured to point to them, you're fine. OR if you are getting the cuda
libs via python dependencies and you are using an "expected" python
dependency structure, you're also fine. The issue for us is that we want
to get the cuda deps via bazel if the user is using bazel (they will be
downloaded regardless of whether the user already has the deps in their
env or not, plus doing it via bazel ensures that the cuda libs are the
versions required by torch). However, the way bazel lays out the nvidia
python deps is not an "expected" layout for the hack in (2). Thus we
implement a patch (based on the pytorch one) that will successfully
find, load, and link the .so files in a bazel env. This patch has to be
imported before torch to work.

That patch is fine for most of our cases, but runs into trouble when it
comes to ray workers. There, we do not control the python script that
gets started, and thus can't ensure that the patch gets imported before
torch. However, in that specific situation we happen to already be
intercepting the python interpreter launch to be a custom script, and
can thus dynamically modify `LD_LIBRARY_PATH` so that it can find the
.so files from nvidia. Why do we need to dynamically modify it instead
of doing it at build time? Bazel restrictivity again 😛. The filesystem
path where bazel places the .so files depends on the specific bazel
target used to build the image, and is thus can't be predicted in
isolation from the specific context of the build. Luckily, the dynamic
modification is very fast, adding a negligible amount of time to the pod
launch.

Testing
-------
- Ran [mnist
remotely](https://dev1.dev-usw2-sematic0.sematic.cloud/runs/41f07b81580c4c5fb1be827e607bfaf8)
including using cuda
- Ran both Ray examples remotely
([1](https://dev1.dev-usw2-sematic0.sematic.cloud/runs/1ad314f620c94508af5fa6ddcb1c6a6f),
[2](https://dev1.dev-usw2-sematic0.sematic.cloud/runs/d6f8a3ac28e24c579413eb873c9e2f7b))
- Ran Cifar Ray example locally
- Ran [testing pipeline
remotely](https://dev1.dev-usw2-sematic0.sematic.cloud/runs/f4d4e5a8d9c04bc286c926396af430f5)

---------

Co-authored-by: Josh Bauer <josh@sematic.dev>