Add prompt/response viz (#897)
Add a new type and visualization for textual prompt/response pairs.
Depending on the use case, the prompt or response may range from a
single word to several paragraphs. These objects may often be displayed
in container types such as lists. This PR implements a suitable display
for both the existing and the new UI.


![PromptResponseCurrentUi](https://github.com/sematic-ai/sematic/assets/7181589/bc8959e9-7fe3-475b-8853-59c1fa9d560c)


![PromptResponseNewUi](https://github.com/sematic-ai/sematic/assets/7181589/204a3275-eda3-4dc7-a9c4-f6df90c3015a)

---------

Co-authored-by: Josh Bauer <josh@sematic.dev>