Add Hugging Face types and visualizations (#920)
Closes #912 

Add python types and visualizations for Hugging Face data types. Two of
the types are references to objects on Hugging Face hub (models and
datasets). These each can optionally refer to a specific commit of the
resource on the hub. The visualizations render a "slug" describing what
is being linked to, and an appropriate link. These types have some
optional fields, depending on which model/dataset is being referenced
and whether a specific subset of a dataset is being referenced.

The other type is a reference to a model stored in external storage
(right now, only local storage).

For the new UI, the reference types and the model storage have "short"
and "full" versions, as the full version wouldn't fit comfortably in the
new UI when the objects are members of another structure.


![HfTypesCurrentUi](https://github.com/sematic-ai/sematic/assets/7181589/9c519df6-81e1-4384-961a-cc65f406f7b0)


![HfTypesNewUi](https://github.com/sematic-ai/sematic/assets/7181589/981a588b-8c5c-42cc-9b06-9bba6999127f)

---------

Co-authored-by: Josh Bauer <josh@sematic.dev>