Update deploy.md (#941)
Handling SSL at the python server layer is no longer supported. I don't
imagine that's much of an issue; a lot of people who'd be deploying this
for shared use would be able to set up a reverse proxy. If we need to
add it back, we can file an issue for that, but we should at least keep
the docs accurate until we do that.