Make Docker image URI parsing more robust (#956)
If a docker image URI is surrounded in quotes, it can throw off the
image pushing. This change ensures that we strip quotes if they're
present. This PR also introduces two changes which were introduced to
allow testing the change on a Mac:

- make sure that when a Sematic developer does `bazel run
//sematic/cli:main`, it runs from the directory you ran that command
from (basically as if you had actually executed the `sematic` cli).
- add a build yaml for the add pipeline

Testing
--------

`bazel run //sematic/cli:main -- run --build
./sematic/examples/add/__main__.py -- --cloud --detach`


[run](https://josh.dev-usw2-sematic0.sematic.cloud/runs/97dd47094bb34609961ae9a58fb53e3f#tab=output)