Peft from remote base (#945)
For GPT-J, it's preferable to re-load the base model from HF instead of
using the saved version from the fine tuning. Why? Because fine-tuning
requires loading the model in lower precision, but it's better to do the
eval using the full precision model. This pattern is actually
recommended for fine-tuning in some HF docs. This PR enables our "stored
model" abstraction to optionally re-pull the image from HF, with
different configurations for how it should be pulled and loaded into
memory.

Testing
-------

Have been using with the GPT-J pipeline, and have confirmed it still
works with Flan.

---------

Co-authored-by: Josh Bauer <josh@sematic.dev>