Add GPT-J support to summarization example (#946)
Note: #945 should be merged before this.

This PR adds GPT-J to the summarization fine-tuning example. A lot of
the code changes are related to generalizing to a Causal LM, but also to
help accommodate a need for loading the model in different precisions
for training & eval.

Probably could greatly improve results by playing with hyperparameters.
One example run using GPT-J is
[here](https://josh.dev-usw2-sematic0.sematic.cloud/runs/d9b667aead68438ba9eac9ddc472ed7f)

Testing
-------
- Built the wheel, installed from it, and ran the pipeline with small
inputs for a Flan variant and GPT J on the AWS G5 machine
- Performed the same from my Macbook, but only testing Flan (GPT-J is
too large, & effectively requires GPU support to load in 8 bit for
training).

---------

Co-authored-by: Josh Bauer <josh@sematic.dev>