Improve error handling for unschedulable resolutions (#973)
When resolutions can't be scheduled (ex: because their images use an
outdated version of the API), we currently only record the error in the
server logs, and the resolution appears to be permanently "running" in
the dashboard. This PR fixes it so we fail the resolution and display an
appropriate message in the dashboard.

Ex: re-running an old resolution gave this message in the UI for the new
(automatically failed) run:

> Failed because the runner job could not be scheduled: The resolution
eaa50e87e487423e87acb2c668ebb3cd uses Sematic version 0.28.1, but the
server requires at least version 0.30.0


https://dev1.dev-usw2-sematic0.sematic.cloud/runs/eaa50e87e487423e87acb2c668ebb3cd