Add "runners" and migrate client-side code to use them (#964)
This PR creates the `Runner` interface as a part of the rename of
`Resolver` -> `Runner`. For runners, there is no `.resolve` method, but
there is a `.run` method. Futures don't have a `future.run` method,
rather we are using this as an opportunity to force people towards
`runner.run(future)`. This PR preserves the old resolver classes so that
end-user code will not be immediately broken, but will rather result in
warnings until they migrate to the new structure.

Additionally, all client side code is changed to support/use the
`Runner` classes, terminology, and invocation style (including
'resolution' -> 'pipeline run'). This includes examples. Documentation
has not yet been updated, but will be in a follow-up PR. What is meant
by "all client side?" Which things are NOT yet migrated to new
terminology? There is a summary given in `resolution.py`, but the basic
principle is to allow any code that doesn't execute in the server to use
the new terminology and style. This ensures that users can start moving
to the new structure, but the server can remain agnostic to whether they
have migrated or not, enabling an automatic backwards compatibility for
now. Migrations of the API endpoints, db query code, db tables/ORM
models/fields will all be done in separate PRs.

Testing
-------

Before moving the tests for the runners to be on the runners themselves,
I made sure they still passed on the resolvers (which now inherit from
their corresponding runners). Additionally, I left the examples using
the `Resolver` variants until the end, and ran several of them to ensure
they still worked before migrating them to `Runner`. That should
illustrate that end-user code that has not yet migrated to `Runner`
should be fine.

After migrating the examples to the new format, I tested several of them
(TODO: haven't actually done this yet).

Also:
- tested runner failure for all runners
- tested user function failure for all runners
- tested ray integration locally/in the cloud
- tested CLI cancellation
- checked cleaner logs