Add migration to bring Postgres in parity with SQLAlchemy (#903)
this migration & related code sqlalchemy model changes should bring our
postgres database schema in sync with the schema defined in sqlalchemy.
we need this so that we can drop our in-house migration framework and
switch to alembic, which can automatically generate migrations for basic
schema changes and enforce correctness.

there's quite a bit of funny business in here, so i'll add comments to
explain individual changes. please review carefully!

i still need to test with sqlite better, so there might be some hiccups
with that.

there are some risks associated with rolling out the contents of this pr
to customers, because we're trusting that several of the
newly-non-nullable columns are actually that way, i.e. they don't
actually contain null values. as far as i can see in our dbs, this is
the case, but it's hard to be sure for customers since postgres did not
enforce these constraints thus far.

### testing

#### version support

sqlite3 versions before 3.26.0 don't migrate foreign keys when a table
gets renamed (with our configuration setup). some table alter support
was added to sqlite3 in a subsequent version, but we know from customer
reports that we can't assume that this exists either. i installed the
oldest possible python3 version that we officially support (3.8.x) on an
m1 mac (3.8.10) and ubuntu 22.04 (3.8.0). of these two, the oldest
version of sqlite3 seems to be 3.37.2, which is newer than than 3.26.0.

this means that for sqlite3 we _can_ assume that foreign keys _will_
follow a table when the table is renamed, but we cannot assume that
tables can be altered in any way. postgres should have no concerns for
the features used in this pr.

#### sqlite data

tested against some reference sqlite snapshots we have from 0.17.2,
0.18.1, and 0.27.0, helpfully provided by @tscurtu! also tested against
a brand new db with no data/schema. all succeed and the server loads
with no errors.

#### postgres data

tested against my own latest 0.31.2 snapshot.

resolves #302, as well as most instances (but not all) of #303.