Add the Grafana panel to new dashboard as a plugin (#976)
Apart from adding the Grafana panel, this PR also introduced the plugin
system explained below:

1. Bundled the
[`react-slot-fill`](https://github.com/camwest/react-slot-fill) library
into the code base as source code. I included the source code instead of
consuming an NPM dependency because the current version of
`react-slot-fill` does not support React 18. The source code is copied
over, structure-adjusted, API upgraded, and trouble-shot to fix any
compatibility issues with React 18. Eventually, it works! In the future,
we can consider releasing this miraculous library as a standalone NPM
package to share with the community. At that point, the source code will
be replaced with the consumption of the NPM package.

2. The plugin system heavily leverages the mechanism of
`react-slot-fill`.

As a high-level view, to get the plugin system working, 

Step 1, in `RunTabs.tsx` we define a `<Slot>`
```
</TabContext>
    <TabPanel value="input"/>
    <TabPanel value="output"/>
    ....
    <Slot name="run-tabs" />
</TabContext>
```

Step 2, in `main` package's `index.tsx`, we render a
```
<PluginsLoader />
```

Step 3, in `PluginsLoader.tsx`, we do
```
import GrafanaPlugin from "@sematic/addon-grafana";

export default function PluginsLoader() {
    return <GrafanaPlugin/>;
}
```

The `<GrafanaPlugin/>` renders a `<Fill>` like this:

```
return <Fill name="run-tabs" >
        <TabPanel value="Grafana">
            <GraphanaPanel />
        </TabPanel>
    </Fill>
```

So whenever and wherever the `<GrafanaPlugin />` is rendered, its
content under `<Fill>`will be attached to the `<Slot>` with the same
name. Therefore, we expand the panels the `RunTabs.tsx` will render.

The `<Fill>` and `<Slot>` can live at dispersed locations in the React
Virtual DOM tree as long as they have a common ancestor node of
`<SlotFillProvider>`. This makes the implementation decoupled. Entirely
suitable for the plugin's system.

3. Later, for whatever reason, we need to purge the Grafana plugin,
deleting the entire "@sematic/addon-grafana" package and removing the
reference from `PluginsLoader.tsx` are all we need to do.

4. The end product of the Grafana panel looks like this:

<img width="1354" alt="image"
src="https://github.com/sematic-ai/sematic/assets/133257643/23f9bd59-892c-4796-b10b-59b112a0fe5e">

The iframe cannot render due to mocked data, and the target Grafana
deployment does not allow cross-origin. It can be seen from the iframe
attribute that the `title` and `src` are generated as expected.