[2/N] Add Organizations associations (#917)
This is the second of a series of PRs that will introduce the capability
of defining and managing user organizations. The previous PR in the
chain is #915.

This PR propagates the associations between existing data models and the
`organizations` model, without code propagations. The new foreign key
columns remain `NULL`. These will start to be populated once the
Resolution submission API will supply an Organization ID.