Enforce Runs fields selection (#1005)
The Server API endpoint that fetches lists of Runs has a parameter that
can restrict the payload content to a subset of fields, instead of
returning the entire Run objects. This feature is currently restricted
to only Run ids.

The enforcement only functions correctly when garbage collection
parameters are specified, and does not function on the general case.

This PR corrects the behavior and adds a unit test for validation.