Improve local cancellation behavior (#1004)
Closes #282, #855, #983

This PR began as a way to address #983, which shows that local runs can
be left stuck in a "running" state in the DB even after being cancelled
by a CTRL+C. Such a behavior would pollute user's local DBs and future
run searches that search by status. There is no convenient way to move
these runs into a terminal state from the UI once they get stuck in that
state (though using the cleaner CLI with `--force` should do it). This
PR attempted to address this by making the cancellation flow more
explicit so we have more control over how it is handled. The result of
doing this was that a number of other cancellation-related behaviors
were improved as well (even #984, which we had not intended to
improve!).

Here's the basic idea. There were two "entrypoints" to the runner's
cancellation logic: one via OS signal, and one via SocketIO event. This
PR makes this flow less duplicated by having the SocketIO cancellation
handler send a SIGINT to the runner process so that the handler code
there will always be used. The next step in making the flow more
explicit was to have the signal handler raise a `CancellationError`.
Since signal handlers, inline runs, and the runner loop all run on the
main thread, this will stop the runner loop (and as a nice side-effect,
the inline runs as well). `CancellationError` is also specially handled
such that it never appears as a failure, which means that (at least for
the local runner) we should stop seeing situations where parts of
canceled pipelines get marked as failed instead.

Testing
-------
- executed the mnist pipeline locally and cancelled with ctrl+c during
the training step. Did this multiple times and confirmed that the runs
that hadn't completed were all marked as canceled.
- did the same thing, but canceled via the UI instead of ctrl+c
- executed the testing pipeline using the cloud runner and cancelled via
the UI several times. Ensured it still moved all runs to terminal states
as expected. As a side note, the frequency of "canceled runs appearing
as failed" did appear to decrease for the cloud runner as well, but
doesn't seem to have been 100% fixed