Fix a bug and pin libs for fine-tuning (#1024)
It turns out that some library versions that have been released since
this was created have broken some functionality. This pins the libs to
working versions. There was also a bug introduced in the gradio refactor
that made the prompt get truncated to only the first character. This
fixes that.