Make runner reentrant (#1037)
The cloud runner is sometimes restarted unexpectedly by K8s. Prior to
this PR, this was handled by starting an entirely new pipeline run, with
a new id, starting where the old one left off. This PR changes the
behavior such that the runner is instead able to re-create its internal
state and continue from where it left off. This (unlike the previous
behavior) should be transparent to end-users.

Testing
-------

Disabled the signal handlers in the StateMachineRunner, so it wouldn't
interpret `kubectl delete pod ...` as cancellations, then performed the
following tests, using `kubectl delete pod` on the runner pod to emulate
evictions:

- With an entirely inline graph, interrupted mid-execution.
[Run](https://josh.dev-usw2-sematic0.sematic.cloud/runs/390bf91a57524b38acff77df6265fce7#run=e6d3cc3d90dc445d920c129ff8663be7&tab=source)
- With standalone functions, interrupted mid-execution.
[Run](https://josh.dev-usw2-sematic0.sematic.cloud/runs/41e720c95c95449f8e74a5f91e25a0ea#run=741f1d1d261446be88af714a9411a3ed&tab=source)
- With multiple base images, interrupted mid-execution.
[Run](https://josh.dev-usw2-sematic0.sematic.cloud/runs/41e720c95c95449f8e74a5f91e25a0ea#run=741f1d1d261446be88af714a9411a3ed&tab=source)
- With an implicit make_list, interrupted mid-execution.
[Run](https://josh.dev-usw2-sematic0.sematic.cloud/runs/d26de86afc9b401ba596690918db1b78)
- Validated that rerun-from-here still works.
[Run](https://josh.dev-usw2-sematic0.sematic.cloud/runs/9034aa1bc6eb4e238d8477d8a172a7a7)

---------

Co-authored-by: Josh Bauer <josh@sematic.dev>