Make image macro configurable (#1041)
Some users want to be able to use their own bazel libraries to construct
docker images. This makes that possible.

Testing
--------
Explicitly configured `image_macro` for the testing pipeline and
submitted. Left it at default and submitted. Changed to another macro
that didn't actually build an image and confirmed that submission
failed.

Co-authored-by: Josh Bauer <josh@sematic.dev>