Fix log view regression (#1042)
In the new log view UI, it was waiting to present the logs for a run
until the run reached a state where the logs would be available.
However, while logs appear during the `SCHEDULED` state, while the run
is executing, the UI considered this to be a state where logs were not
available. The effect was that the logs for a run would not show up
until after the run completed. This regression appeared in v0.34.0. This
PR fixes the regression.

Testing
-------
Submitted the testing pipeline including a "sleep" step, and observed
that the logs for the step were shown during the execution of the run.