Improve `sematic version` CLI (#1045)
The `sematic version` CLI was a bit misleading, as it checked the
client's version and called it the server version. This PR changes it so
that both the client and the server version are reported, along with the
min version from the actual server.

Testing
-------

```
$ sematic version
Sematic client v0.34.0 is installed.
The server is at version v0.34.0
Client versions >= v0.30.0 are supported.
Python v3.8.15 is running this binary.
```

And with an unsupported server/client combo:

```
$ sematic version
Sematic client v0.34.0 is installed.
The Sematic API client is at version 0.34.0, which is newer than the Sematic server's version of 0.33.0. Please downgrade your client to 0.33.0, or ask your server admin to upgrade to at least 0.34.0.
```