Have the cancel CLI use the API from the config (#1052)
The CLI for `cancel` was currently always using the local user env
(since that is the default for CLI commands). But really we want it to
communicate with whatever API the user's config is set up to talk to.

Testing
-------

Made sure my Sematic config was pointing to a remote server and used
this command. Ensured it used the remote server to complete the request.