Have the server log more errors & warnings (#1053)
There were a couple of situations where it would be helpful to have
errors logged in the server logs, but that wasn't happening. Basically
(1) anything that was going uncaught was relying on the `uvicorn.*`
logger to get logged, whereas we want it to be handled by the
`sematic.*` logger as well, (2) any time we were explicitly returning a
json error response, we wouldn't log anything to the server logs unless
the particular code calling `jsonify_error` had already logged the
exception. This PR addresses both of these.

Testing
--------

Ran the local server and intentionally injected exceptions, both
uncaught ones and ones that resulted in a `jsonify_error`. Made sure the
exceptions made it to the local server logs. Then did the same thing
with code deployed to my namespace.