Allow custom image pull secrets (#1055)
In some cases, runner and standalone pods may need to pull their images
from a private registry that isn't automatically accessible on the basis
of the service accounts already associated with the pod. This PR makes
it possible to specify custom image pull secrets for these pods.

Testing
-------
Deployed to my namespace and tested with and without configuring
`worker.image_pull_secrets`. Confirmed that the pods got created with
the proper configuration for their imagePullSecrets.