Remove configurability for dedicated socket io (#1060)
You actually get errors connecting to socketio if you don't use a
dedicated socketio. So the default setting is broken. This PR makes it
so that this is no longer configurable, and you must have a dedicated
socketio server. It also changes the default value for whether to create
a service to `true`, which is the more appropriate value (that is
another place where a vanilla install would fail to work as expected).

Tested by deleting the staging deployment, removing the configs for the
standalone SocketIO server and the service creation from the staging
deployment's chart, and installed the deployment to staging again.