Change the default to be usage of the new UI (#1062)
The new dashboard has been vetted well enough that we can change the
default behavior to be using the new UI. We still likely want to keep
the old one around for a few more releases in case this change surfaces
new users that haven't tried the dashboard and run into problems. If
that happens, we can fix their issues and let them manually switch back
to the old one as a mitigation.

Testing
--------
Cleared my local storage, rebuilt & started the app, and refreshed my
page. Confirmed it used the new UI. Used the toggles to switch between
the new and old UIs and confirmed that they still toggle.