Fix edge-case in timeout comparisons (#1096)
There was a bug that if two runs timed out at the same exact time (ex:
parallel runs launched simultaneously), the comparison logic would fall
back to comparing `Future` objects directly, which don't support the `<`
operator. This switches so that in such cases we fall back to comparing
the ids instead of the Future objects.