Fail more loudly if artifact data writes are not successful (#1097)
closes #1089 

When doing a `PUT` to write artifact data, we are currently not raising
an error directly. Likely an error will still occur when something
downstream tries to *use* the artifact data (unless it's the last
artifact in a pipeline...), but we should fail more explicitly when the
error first occurs.